+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [16.0.3] - 2023-03-13
 ### Changes
 
 - Adding refresh tokens to refresh calls even if access token isn't present to make manual testing easier.